requirements.txt
flask==3.0.0
twilio==8.10.0
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
pyresparser==1.0.6
spacy==2.3.5
requests==2.31.0
ngrok==1.1.0
#openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
